+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/rate_limiter.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
 use Symfony\Component\HttpKernel\EventListener\RateLimitAttributeListener;
+use Symfony\Component\RateLimiter\RateLimiterBuilder;
 use Symfony\Component\RateLimiter\RateLimiterFactory;
 
 return static function (ContainerConfigurator $container) {
@@ -31,5 +32,11 @@ return static function (ContainerConfigurator $container) {
         ->set('rate_limiter.attribute_listener', RateLimitAttributeListener::class)
             ->tag('kernel.event_subscriber')
             ->args([tagged_locator('rate_limiter', 'name')])
+
+        ->set('limiter_builder', RateLimiterBuilder::class)
+            ->args([
+                abstract_arg('storage'),
+                null,
+            ])
     ;
 };